+++ lms/lms/doctype/loan_application/loan_application.py
@@ -484,6 +484,8 @@ class LoanApplication(Document):
         lender = self.get_lender()
         self.minimum_sanctioned_limit = lender.minimum_sanctioned_limit
         self.maximum_sanctioned_limit = lender.maximum_sanctioned_limit
+        values1 = {"value": ["{}".format(lender), lms.get_linenumber()]}
+        lms.create_log(values1, "status_pledge_accepted_by_lender")
 
         if (
             self.status == "Approved"
@@ -491,23 +493,56 @@ class LoanApplication(Document):
             and not self.loan_margin_shortfall
             and not self.application_type == "Pledge More"
         ):
+            values2 = {
+                "value": [
+                    self.status,
+                    self.lender_esigned_document,
+                    self.loan_margin_shortfall,
+                    self.application_type,
+                    lms.get_linenumber(),
+                ]
+            }
+            lms.create_log(values2, "status_pledge_accepted_by_lender ")
             frappe.throw("Please upload Lender Esigned Document")
         elif self.status == "Approved":
+            values3 = {"value": [self.status, lms.get_linenumber()]}
+            lms.create_log(values3, "status_pledge_accepted_by_lender ")
             current = frappe.utils.now_datetime()
             expiry = frappe.utils.add_years(current, 1) - timedelta(days=1)
             self.expiry_date = datetime.strftime(expiry, "%Y-%m-%d")
         elif self.status == "Pledge accepted by Lender":
+            values4 = {"value": [self.status, lms.get_linenumber()]}
+            lms.create_log(values4, "status_pledge_accepted_by_lender")
+
             if self.pledge_status == "Failure":
+                values5 = {"value": [self.pledge_status, lms.get_linenumber()]}
+                lms.create_log(values5, "status_pledge_accepted_by_lender")
+
                 frappe.throw("Sorry! Pledge for this Loan Application is failed.")
 
             total_approved = 0
             total_collateral_value = 0
 
+            values6 = {"value": [lms.get_linenumber()]}
+            lms.create_log(values6, "status_pledge_accepted_by_lender")
+
             for i in self.items:
+                values7 = {"value": [i, lms.get_linenumber()]}
+                lms.create_log(values7, "status_pledge_accepted_by_lender")
+
                 if i.get("pledge_status") == "Failure" and i.lender_approval_status in [
                     "Approved",
                     "Rejected",
                 ]:
+                    values8 = {
+                        "value": [
+                            i.get("pledge_status"),
+                            (i.lender_approval_status),
+                            lms.get_linenumber(),
+                        ]
+                    }
+                    lms.create_log(values8, "status_pledge_accepted_by_lender")
+
                     frappe.throw(
                         "Pledge failed for ISIN - {}, can't Approve or Reject".format(
                             i.isin
@@ -515,7 +550,23 @@ class LoanApplication(Document):
                     )
 
                 elif i.get("pledge_status") == "Success":
+                    values9 = {
+                        "value": [
+                            i.get("pledge_status"),
+                            lms.get_linenumber(),
+                        ]
+                    }
+                    lms.create_log(values9, "status_pledge_accepted_by_lender")
+
                     if i.lender_approval_status == "Pledge Failure":
+                        values10 = {
+                            "value": [
+                                (i.lender_approval_status),
+                                lms.get_linenumber(),
+                            ]
+                        }
+                        lms.create_log(values10, "status_pledge_accepted_by_lender")
+
                         frappe.throw(
                             "Already pledge success for {}, not allowed to set Pledge Failure.".format(
                                 i.isin
@@ -523,13 +574,31 @@ class LoanApplication(Document):
                         )
 
                     elif i.lender_approval_status == "":
+                        values11 = {
+                            "value": [
+                                (i.lender_approval_status),
+                                lms.get_linenumber(),
+                            ]
+                        }
+                        lms.create_log(values11, "status_pledge_accepted_by_lender")
+
                         frappe.throw("Please Approve/Reject {}".format(i.isin))
 
                     if i.lender_approval_status == "Approved":
+                        values31 = {
+                            "value": [
+                                (i.lender_approval_status),
+                                lms.get_linenumber(),
+                            ]
+                        }
+                        lms.create_log(values31, "status_pledge_accepted_by_lender")
                         total_approved += 1
                         total_collateral_value += i.amount
 
             if total_approved == 0:
+                values12 = {"value": [(total_approved), lms.get_linenumber()]}
+                lms.create_log(values12, "status_pledge_accepted_by_lender")
+
                 frappe.throw(
                     "Please Approve atleast one item or Reject the Loan Application"
                 )
@@ -594,19 +663,52 @@ class LoanApplication(Document):
                 )
 
             if self.application_type in ["New Loan", "Increase Loan"]:
+                values13 = {"value": [(self.application_type), lms.get_linenumber()]}
+                lms.create_log(values13, "status_pledge_accepted_by_lender")
+
                 if drawing_power < self.minimum_sanctioned_limit:
+                    values14 = {
+                        "value": [
+                            drawing_power,
+                            self.minimum_sanctioned_limit,
+                            lms.get_linenumber(),
+                        ]
+                    }
+                    lms.create_log(values14, "status_pledge_accepted_by_lender")
+
                     frappe.throw(
                         "Sorry! This Loan Application can not be Approved as its Drawing power is less than Minimum Sanctioned Limit."
                     )
 
         if self.status == "Pledge executed":
             total_collateral_value = 0
+            values15 = {"value": [self.status, lms.get_linenumber()]}
+            lms.create_log(values15, "status_pledge_accepted_by_lender")
             for i in self.items:
+                values16 = {"value": [i, lms.get_linenumber()]}
+                lms.create_log(values16, "status_pledge_accepted_by_lender")
+
                 if i.pledge_status == "Success" or i.pledge_status == "":
+                    values17 = {
+                        "value": [
+                            i.pledge_status,
+                            lms.get_linenumber(),
+                        ]
+                    }
+                    lms.create_log(values17, "status_pledge_accepted_by_lender")
+
                     if (
                         i.lender_approval_status == "Approved"
                         or i.lender_approval_status == ""
                     ):
+                        values18 = {
+                            "value": [
+                                i.lender_approval_status,
+                                lms.get_linenumber(),
+                            ]
+                        }
+                        lms.create_log(values18, "status_pledge_accepted_by_lender")
+
                         total_collateral_value += i.amount
                         self.total_collateral_value = round(total_collateral_value, 2)
                         if self.instrument_type == "Shares":
@@ -636,6 +738,15 @@ class LoanApplication(Document):
                             if drawing_power < self.maximum_sanctioned_limit
                             else self.maximum_sanctioned_limit
                         )
+                        values19 = {
+                            "value": [
+                                total_collateral_value,
+                                self.total_collateral_value,
+                                self.drawing_power,
+                                lms.get_linenumber(),
+                            ]
+                        }
+                        lms.create_log(values19, "status_pledge_accepted_by_lender")
 
                     # elif (
                     #     i.lender_approval_status == "Rejected"
@@ -656,7 +767,12 @@ class LoanApplication(Document):
 
         # On loan application rejection mark lender approvel status as rejected in loan application items
         if self.status == "Rejected":
+            values20 = {"value": [self.status, lms.get_linenumber()]}
+            lms.create_log(values20, "status_pledge_accepted_by_lender")
             for i in self.items:
+                values21 = {"value": [i, lms.get_linenumber()]}
+                lms.create_log(values21, "status_pledge_accepted_by_lender line354")
+
                 i.lender_approval_status = "Rejected"
 
         self.total_collateral_value_str = lms.amount_formatter(
@@ -668,10 +784,17 @@ class LoanApplication(Document):
         )
 
     def on_update(self):
+        print("Prod")
         if self.status == "Approved":
             if not self.loan:
+                values21 = {"value": [self.status, lms.get_linenumber()]}
+                lms.create_log(values21, "status_pledge_accepted_by_lender")
+
                 loan = self.create_loan()
             else:
+                values21 = {"value": [(self.status, lms.get_linenumber())]}
+                lms.create_log(values21, "status_pledge_accepted_by_lender")
+
                 loan = self.update_existing_loan()
             frappe.db.commit()
             if self.application_type in ["New Loan", "Increase Loan"]:
@@ -680,17 +803,38 @@ class LoanApplication(Document):
 
             if not self.loan:
                 # new loan agreement mapping
+                values22 = {"value": [self.loan, lms.get_linenumber()]}
+                lms.create_log(values22, "status_pledge_accepted_by_lender")
+
                 self.map_loan_agreement_file(loan)
             elif (
                 self.loan
                 and self.lender_esigned_document
                 and not self.loan_margin_shortfall
             ):
+                values23 = {
+                    "value": [
+                        self.loan,
+                        self.lender_esigned_document,
+                        self.loan_margin_shortfall,
+                        lms.get_linenumber(),
+                    ]
+                }
+                lms.create_log(values23, "status_pledge_accepted_by_lender")
+
                 # increase loan agreement mapping
                 self.map_loan_agreement_file(loan, increase_loan=True)
 
             if self.loan_margin_shortfall:
                 # if shortfall is not recoverd then margin shortfall status will change from request pending to pending
+                values24 = {
+                    "value": [
+                        self.loan_margin_shortfall,
+                        lms.get_linenumber(),
+                    ]
+                }
+                lms.create_log(values24, "status_pledge_accepted_by_lender")
+
                 loan_margin_shortfall = frappe.get_doc(
                     "Loan Margin Shortfall", self.loan_margin_shortfall
                 )
@@ -726,6 +870,18 @@ class LoanApplication(Document):
                         "loan_margin_shortfall": loan_margin_shortfall.name,
                     },
                 )
+
+                values25 = {
+                    "value": [
+                        loan_margin_shortfall,
+                        under_process_la,
+                        pending_loan_transaction,
+                        pending_sell_collateral_application,
+                        lms.get_linenumber(),
+                    ]
+                }
+                lms.create_log(values25, "status_pledge_accepted_by_lender")
+
                 if (
                     (
                         not pending_loan_transaction
@@ -735,6 +891,17 @@ class LoanApplication(Document):
                     and loan_margin_shortfall.status == "Request Pending"
                     and loan_margin_shortfall.shortfall_percentage > 0
                 ):
+                    values26 = {
+                        "value": [
+                            loan_margin_shortfall,
+                            under_process_la,
+                            pending_loan_transaction,
+                            pending_sell_collateral_application,
+                            lms.get_linenumber(),
+                        ]
+                    }
+                    lms.create_log(values26, "status_pledge_accepted_by_lender")
+
                     loan_margin_shortfall.status = "Pending"
                     loan_margin_shortfall.save(ignore_permissions=True)
                     frappe.db.commit()
@@ -742,21 +909,61 @@ class LoanApplication(Document):
         elif self.status == "Pledge accepted by Lender":
             approved_isin_list = []
             rejected_isin_list = []
+            values27 = {"value": [self.status, lms.get_linenumber()]}
+            lms.create_log(values27, "status_pledge_accepted_by_lender")
             for i in self.items:
+                values27 = {"value": [i, lms.get_linenumber()]}
+                lms.create_log(values27, "status_pledge_accepted_by_lender")
+
                 if i.lender_approval_status == "Approved":
+                    values28 = {
+                        "value": [
+                            i.lender_approval_status,
+                            lms.get_linenumber(),
+                        ]
+                    }
+                    lms.create_log(values28, "status_pledge_accepted_by_lender")
+
                     approved_isin_list.append(i.isin)
                 elif i.lender_approval_status == "Rejected":
+                    values28 = {
+                        "value": [
+                            i.lender_approval_status,
+                            lms.get_linenumber(),
+                        ]
+                    }
+                    lms.create_log(values28, "status_pledge_accepted_by_lender")
+
                     rejected_isin_list.append(i.isin)
 
             if len(approved_isin_list) > 0:
+                values29 = {"value": [approved_isin_list, lms.get_linenumber()]}
+                lms.create_log(values29, "status_pledge_accepted_by_lender")
+
                 self.update_collateral_ledger(
                     {"lender_approval_status": "Approved"},
                     "application_doctype = 'Loan Application' and application_name = '{}' and isin IN {}".format(
                         self.name, lms.convert_list_to_tuple_string(approved_isin_list)
                     ),
                 )
+                val590 = {
+                    "value": [
+                        approved_isin_list,
+                        lms.get_linenumber(),
+                        "return from update collateral list",
+                    ]
+                }
+                lms.create_log(val590, "status_pledge_accepted_by_lender")
 
             if len(rejected_isin_list) > 0:
+                values29 = {
+                    "value": [
+                        rejected_isin_list,
+                        lms.get_linenumber(),
+                    ]
+                }
+                lms.create_log(values29, "status_pledge_accepted_by_lender")
+
                 self.update_collateral_ledger(
                     {"lender_approval_status": "Rejected"},
                     "application_doctype = 'Loan Application' and application_name = '{}' and isin IN {}".format(
@@ -771,6 +978,8 @@ class LoanApplication(Document):
                     },
                     tokens=lms.get_firebase_tokens(self.get_customer().user),
                 )
+                values30 = {"value": [fa, lms.get_linenumber()]}
+                lms.create_log(values30, "status_pledge_accepted_by_lender")
             except Exception:
                 pass
             finally:
@@ -814,6 +1023,16 @@ class LoanApplication(Document):
                         "loan_margin_shortfall": loan_margin_shortfall.name,
                     },
                 )
+                values30 = {
+                    "value": [
+                        loan_margin_shortfall,
+                        pending_sell_collateral_application,
+                        pending_loan_transaction,
+                        lms.get_linenumber(),
+                    ]
+                }
+                lms.create_log(values30, "status_pledge_accepted_by_lender")
+
                 if (
                     (
                         not pending_loan_transaction
@@ -828,14 +1047,31 @@ class LoanApplication(Document):
                     frappe.db.commit()
 
             if not self.loan and not self.loan_margin_shortfall:
+                values30 = {"value": ["True", lms.get_linenumber()]}
+                lms.create_log(values30, "status_pledge_accepted_by_lender")
+
                 customer = self.get_customer()
                 if customer.pledge_securities:
                     customer.pledge_securities = 0
                     customer.save(ignore_permissions=True)
                     frappe.db.commit()
+                    values30 = {
+                        "value": [
+                            customer.pledge_securities,
+                            lms.get_linenumber(),
+                        ]
+                    }
+                    lms.create_log(values30, "status_pledge_accepted_by_lender")
 
             # On loan application rejection mark lender approvel status as rejected in collateral ledger as well 23-09-2021 Poonam
             loan_application_isin_list = [i.isin for i in self.items]
+            values30 = {
+                "value": [
+                    loan_application_isin_list,
+                    lms.get_linenumber(),
+                ]
+            }
+            lms.create_log(values30, "status_pledge_accepted_by_lender")
 
             self.update_collateral_ledger(
                 {"lender_approval_status": "Rejected"},
@@ -1378,6 +1614,7 @@ class LoanApplication(Document):
                 "PledgeeBOID": self.pledgee_boid,
                 "PRFNumber": prf_number,
                 "ExpiryDate": self.expiry_date.strftime("%d%m%Y"),
+                "ReasonCode": "06",
                 "ISINDTLS": securities_array,
             }
             headers = las_settings.cdsl_headers()
@@ -1844,7 +2081,6 @@ def process_pledge(loan_application_name=""):
 
     current_hour = int(utils.nowtime().split(":")[0])
     las_settings = frappe.get_single("LAS Settings")
-
     if (
         las_settings.scheduler_from_time
         <= current_hour
@@ -1856,7 +2092,6 @@ def process_pledge(loan_application_name=""):
             fields=["count(name) as count", "status"],
             filters={"status": "Executing pledge", "instrument_type": "Shares"},
         )
-
         if is_pledge_executing[0].count == 0:
             filters_query = {
                 "status": "Waiting to be pledged",
@@ -1864,7 +2099,6 @@ def process_pledge(loan_application_name=""):
             }
             if loan_application_name:
                 filters_query["name"] = loan_application_name
-
             loan_application = frappe.get_all(
                 "Loan Application",
                 fields="name, creation",
@@ -1894,11 +2128,13 @@ def only_pdf_upload(doc, method):
 
 @frappe.whitelist()
 def actions_on_isin(loan_application):
+    print("Inside Actions on isin")
     loan_application = json.loads(loan_application)
     loan_application_doc = frappe.get_doc("Loan Application", loan_application["name"])
     if loan_application_doc.status == "Pledge executed":
         total_collateral_value = 0
         drawing_power = 0
+        print("inside internal if")
         for i in loan_application["items"]:
             if i["pledge_status"] == "Success" or i["pledge_status"] == "":
                 if (
@@ -1955,5 +2191,5 @@ def actions_on_isin(loan_application):
                 loan_application["pledged_total_collateral_value"]
             ),
         }
-
+        lms.create_log(response, "status_pledge_accepted_by_lender")
         return response