+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [v5.5.1-uat]
+
+## Added
+
+- [Feat: sanction date field added](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/528)
+
 ## [v5.5.0-uat]
 
 ## Added
@@ -95,17 +101,23 @@
 - [PII changes](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/510)
 - [API Doc changes](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/491)
 
-## [v3.2.0-uat]
+## [v3.3.0-uat]
 
 ### Added
 
 - [Feat: Mutual Fund Implementation](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/489)
 
-## [v3.1.0]
+## [v3.2.0] - 2022-09-16
 
 ### Added
 
-- [Feat : Feat : OTP SMS Template Change](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/527)
+- [Feat: Spark app version doctype and api changes, Payment auto approval, otp on email, ReasonCode=06 changes and Email Campaign](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/518)
+
+## [v3.1.0] - 2022-09-01
+
+### Added
+
+- [Feat : OTP SMS Template Change](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/527)
 
 ## [v3.0.1] - 2022-03-08